pyproject.toml
[project]
name = "ad"
version = "0.1.0"
description = ""
authors = [
    {name = "babou",email = "bastien.angeletti@ynov.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ldap3 (==2.9.1)",
    "python-dotenv (==1.1.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
